ProPay einbinden double zu int geaendert (paatty32 & AndrejBri)

+++ Elite-Verleih/src/main/java/de/hhu/abschlussprojektmeangirlzhackstreetboys/EliteVerleih/controller/ProfilBearbeitenController.java
@@ -25,7 +25,7 @@ public class ProfilBearbeitenController {
     		return "redirect:/";
     	}
     	Benutzer benutzer = benutzerManager.getBenutzerById(id);
-        double geld = sync.getAccount(benutzer.getBenutzerName()).getAmount();
+        int geld = (int) sync.getAccount(benutzer.getBenutzerName()).getAmount();
 
         model.addAttribute("Betrag", geld);
     	model.addAttribute("benutzer",benutzer);
